master: Report changelog application failures in shadow master
Add a new LIZ_MLTOMA_CHANGELOG_APPLY_ERROR message that is sent by
shadow masters to notify the master server about changelog application
problems.

Upon reception of such a message, master dumps metadata and replies. If
the message is about checksum error, the master server also rechecks its
checksum before reply. After this operation shadow master removes its
metadata and downloads a fresh copy from the master server.

+++ src/metarestore/restore.h
@@ -25,5 +25,5 @@
 enum class RestoreRigor { kIgnoreParseErrors, kDontIgnoreAnyErrors };
 
 void restore_reset();
-int restore(const char* filename, uint64_t lv, const char* ptr, RestoreRigor rigor);
+uint8_t restore(const char* filename, uint64_t lv, const char* ptr, RestoreRigor rigor);
 void restore_setverblevel(uint8_t _vlevel);
